requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
click==8.0.1; python_version >= '3.6'
flask==2.0.1
itsdangerous==2.0.1; python_version >= '3.6'
jinja2==3.0.1; python_version >= '3.6'
markupsafe==2.0.1; python_version >= '3.6'
marshmallow==3.13.0
python-decouple==3.4
werkzeug==2.0.1; python_version >= '3.6'
